remove needless module
